+++ CHANGELOG.md
@@ -273,3 +273,4 @@
 # 0.3.2
 * Define experimental methods: `->>=`, `>>=-`, `-=<<`, `=<<-`
 * Define `interruptable` method for `Continuation`
+* Define `Catchable` typeclass to catch errors from `Conclusion`